requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic>=2.7.0
openai>=1.0.0
streamlit>=1.36.0
requests>=2.32.0
python-dotenv>=1.0.1
pypdf>=5.0.0
python-docx>=1.1.2
